requirements.txt
Pyrogram==2.0.106
yt-dlp==2024.5.27
APScheduler==3.10.4
asyncio==3.4.3
TgCrypto==1.2.5
Pillow==10.3.0